Minor bug Fixes, New Admin Command, and Reformatting - Moved Commands from enumys to the commands package - Added ADMIN command to view or directly set a players Admin Mode - Changed default ADMIN_TOGGLED message from `now` mto `currently`. I think this makes more sense when the command is used both after changing and for viewing - Changed default value of adminEnabled to true. Still requires ADMIN permissions so having the value true for non admin players wont do anything - Removed Freeze and replaced with a `s != null` check on sign update. This fixes signs made with commands not updating and I believe still fixes what freeze was added to fix as well.
